+++ src/components/App/Header/HeaderTop/Top.module.css
@@ -17,10 +17,3 @@
     display: none;
   }
 }
-
-.inner {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-}
