+++ src/components/App/Header/HeaderTop/index.ts
@@ -1,0 +1,1 @@
+export { default } from './HeaderTop';
